requirements.txt
# ───── Core API and Frameworks ─────
fastapi==0.104.1
uvicorn[standard]==0.24.0
aiofiles==23.2.1
pydantic==2.5.0
python-multipart==0.0.6
python-dotenv==1.0.0

# ───── Database & Auth ─────
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0

# ───── ML / AI Stack ─────
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
open-clip-torch==2.24.0
transformers>=4.35.0
scikit-learn>=1.0
scikit-image>=0.22.0
numpy>=1.24.0
pillow>=10.0.0
colorthief>=0.2.1
tqdm>=4.66.0

# ───── Vector DB & Utilities ─────
weaviate-client==4.6.2
sqlite-vec>=0.1.0

# ───── Optional / Misc ─────
opencv-python-headless>=4.8.0
